Update Docker file and deploy workflow
Changed the way tasks.txt file is mounted into the Docker container in the deploy.yml file. Also removed the unnecessary copying of tasks.txt in the Dockerfile. Now, the text file is not being copied into the Docker image during the build, instead it is mounted from the current directory during the container runtime.